Refactor. slide클래스 분리 후 hot-deal-slide에 적용
